requirements.txt
fastapi
uvicorn
websockets
aiohttp
logging
python-dotenv